backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.10.0
langchain-groq==0.2.0
